Fixes isolated tests for studies dispatcher
